Move packages storage/{table,token} -> storage/target/{table,token}
