+++ internal/pkg/service/buffer/dependencies/dependencies.go
@@ -37,8 +37,8 @@ import (
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/event"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/file"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/statistics"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/storage/table"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/storage/token"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/storage/target/table"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/storage/target/token"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/store"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/store/schema"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/watcher"
