+++ internal/pkg/service/buffer/dependencies/reqproject.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 
 	"github.com/keboola/keboola-as-code/internal/pkg/log"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/storage/table"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/storage/token"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/storage/target/table"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/storage/target/token"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/common/dependencies"
 )
